+++ src/main/java/com/cramirez/customcollections/App.java
@@ -1,0 +1,12 @@
+package com.cramirez.customcollections;
+
+/**
+ * Hello world!
+ *
+ */
+public class App {
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
