+++ src/main/java/com/cramirez/customcollections/list/Iterator.java
