+++ src/main/java/com/cramirez/customcollections/list/linkedlist/Node.java
@@ -1,0 +1,11 @@
+package com.cramirez.customcollections.list.linkedlist;
+
+class Node<T> {
+    Node<T> next;
+    Node<T> previous;
+    T data;
+
+    public Node(T data) {
+        this.data = data;
+    }
+}
